split graphQL generation from collection and registration in the whole schema
